requirements.txt
# Core dependencies
pyspark==3.5.0
mlflow==2.8.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data processing
pandas==2.1.3
numpy==1.24.3
delta-spark==3.0.0

# Kafka
kafka-python==2.0.2

# ML/AI
sentence-transformers==2.2.2
torch==2.1.0
scikit-learn==1.3.2

# Vector databases
pinecone-client==2.2.4
chromadb==0.4.15

# LLM
openai==1.3.5
anthropic==0.7.7
langchain==0.0.350

# Monitoring
evidently==0.4.8

# Databricks
databricks-sql-connector==3.0.0
databricks-feature-store==0.3.0

# Airflow (optional, for orchestration)
apache-airflow==2.7.3

# Utilities
python-dotenv==1.0.0
httpx==0.25.1
python-multipart==0.0.6

